docs(date-picker): optimize events demo
