+++ examples/sites/demos/app/date-picker/events.spec.ts
@@ -2,7 +2,7 @@ import { expect, test } from '@playwright/test'
 
 test('[DatePicker] 测试事件', async ({ page }) => {
   page.on('pageerror', (exception) => expect(exception).toBeNull())
-  await page.goto('http://localhost:7130/pc/date-picker/date-picker-events')
+  await page.goto('http://localhost:7130/pc/date-picker/events')
 
   // blur 事件
   await page.getByRole('textbox').nth(1).click()
